Fix BoneAttachment3d not updating warning on external skeleton change.
This patches an issue where the bone attachment 3d would not update its warning after a change in its target skeleton until the project was reloaded or the tree was updated. Now it updates the moment the user changes the external skeleton property. (sorry for tag, not sure how else to send in a patch)