Reverting org.redotengine package to org.godotengine
The package org.godotengine is used both for the Java portion of the editor launched on Android and for the SDK exposed to third party tools and plugins.
The recent change to rename the package to org.redotengine has broken all the binary (and source) dependencies of third party software.
This changes reverts the package declarations to org.godotengine.

However, this can't be merged as is: we first need to identify all the sources that constitute the editor engine on Android, and move those only to org.redotengine. That is necessary to be able to publish the Redot editor on android separately from Godot (and possibly sign applications built with it).