Prevent BonePropertiesEditor crash due to skeleton use-after-release
This fixes an error we caught in the wild via crash reporting.

It is possible for the BonePropertiesEditor to keep a reference
to a skeleton that is deleted behind its back, and then dereference
and crash.

Additionally, this used to remove signals that might not have been
set, so this guards against these warnings as well.

(cherry picked from commit 9c7669d32c075b338d90789e48750c9197ee14e4)